refactor: use pydantic for API
