+++ pyproject.toml
@@ -17,6 +17,7 @@ aiohttp = "^3.11.10"
 tqdm = "^4.67.1 "
 rich = "^13.9.4"
 yarl = "^1.18.3"
+pydantic = "^2.10.4"
 
 [tool.poetry.group.dev.dependencies]
 pre-commit = "^4.0.1"
